+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 ## CHANGELOG
 
-### 02.08.2020
+### 04.27.2021
+
+- Redo the function in Plot for future features.
+- Add convertor tools for quick and efficient usings, like base64 to string and vice versa.
+
+### 02.08.2021
 
 - Minor bug fixes.
 - Reconfigure some functions of OneData and OneSeries for more convinient usings
 
-### 01.23.2020
+### 01.23.2021
 
 - Minor bug fixes.